@@ -1,6 +1,6 @@
 {
   "name": "@aciiverse/generator-aciifx-cli",
-  "version": "0.2.3",
+  "version": "0.2.4",
   "description": "aciiFX generator",
   "main": "generators/index.js",
   "repository": "github:Aciiverse/aciiFX-CLI",
@@ -35,24 +35,24 @@
     "@types/yeoman-assert": "^3.1.4",
     "@types/yeoman-generator": "^5.2.14",
     "@types/yosay": "^2.0.3",
-    "coveralls": "^3.0.7",
-    "eslint": "^6.6.0",
-    "eslint-config-prettier": "^6.6.0",
-    "eslint-config-xo": "^0.27.2",
-    "eslint-plugin-prettier": "^3.1.1",
-    "husky": "^3.0.9",
-    "jest": "^26.1.0",
-    "lint-staged": "^9.4.3",
-    "prettier": "^1.19.1",
+    "coveralls": "^3.1.1",
+    "eslint": "^9.9.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-config-xo": "^0.46.0",
+    "eslint-plugin-prettier": "^5.2.1",
+    "husky": "^9.1.5",
+    "jest": "^29.7.0",
+    "lint-staged": "^15.2.9",
+    "prettier": "^3.3.3",
     "yeoman-assert": "^3.1.1",
-    "yeoman-test": "^1.7.0"
+    "yeoman-test": "^8.3.0"
   },
   "jest": {
     "testEnvironment": "node"
   },
   "dependencies": {
-    "chalk": "^2.1.0",
-    "yeoman-generator": "^3.1.1",
-    "yosay": "^2.0.2"
+    "chalk": "^5.3.0",
+    "yeoman-generator": "^7.3.2",
+    "yosay": "^3.0.0"
   }
 }
